fix double clicking ".session" files on Windows
SciTE sacrifices shell integration to program on the command line. There are workarounds but why not force the workaround on the programmer to begin with. Sending special characters can be done with an external input file - like all the other tools do.

Potential other workarounds if must use C-style string escapes codes on the command line:

https://www.autoitscript.com/forum/topic/202517-starting-from-a-session-shortcut/
https://www.mail-archive.com/scite-interest@lyra.org/msg02380.html